Rearchitect to support multi selection types. Follow best practice.
  * Remove windows.onload event and take optional canvas.
  * Use ES6 modules. Split up code into selection type.
  * Add webpack build and dev steps.

	modified:   package.json
	new file:   src/circle.js
	modified:   src/index.js
	new file:   src/lasso.js
	new file:   src/rectangle.js
