+++ package.json
@@ -4,6 +4,8 @@
   "description": "Paper.js based selection tool.",
   "main": "index.js",
   "scripts": {
+    "dev": "webpack --mode development",
+    "build": "webpack --mode production",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": "git+https://github.com/lsst-epo/paper-select-tool.git",
