+++ src/circle.js
@@ -1,0 +1,6 @@
+'use strict';
+
+export default function circleInit (canvas) {
+  if (!canvas)
+    canvas = document.createElement('canvas');
+}
